Another simplification that doesn't happen in the latest version

+++ Tests/CSharp/ExpressionTests/ExpressionTests.cs
@@ -2096,7 +2096,7 @@ internal static partial class Module1
 {
     public static void Main()
     {
-        short x = true ? 50 : 100;
+        short x = true ? (short)50 : (short)100;
     }
 }
 ");
